+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - MUI v5 Migration - `Components > TableBody` #9082
 - MUI v5 Migration - `Components > TableSortCell` #9082
 - MUI v5 Migration - `Components > TableRow` #9082
+- MUI v5 Migration - `Components > Notice` #9094
 - React Query for Linodes Landing #9062
 
 ## [2023-05-01] - v1.92.0